BestCombination component and custom hook and seperated logic from app
